+++ CHANGES.md
@@ -4,7 +4,7 @@ videojs-record changelog
 2.2.0 - unreleased
 ------------------
 
-- Package library using webpack (#226)
+- Package library using webpack 4 (#226)
 - Add unit tests using karma/webpack/jasmine (#226)
 - Bump required version for videojs-wavesurfer to 2.3.1 for wavesurfer
   destroy fixes